project/starter/requirements.txt
a2a-sdk>=0.3.6
google-adk>=1.14.1
